College Enrollments

A dataset containing information about college enrollments by state in 2015.

college_enrollment

Format

A data frame with 39 rows and 10 variables. The variables are as follows:

school_id

School identification number

name

Name of school

label

School name abbreviation, if applicable

state

State where school is located

city

City where school is located

lat

Latitude of school location

long

Longitude of school location

liberal_arts

Dummy variable for if the school is liberal arts or not

undergrad

Number of undergraduate students enrolled

enrollment

Total number of students enrolled

Source

The data was sourced from QAC251 - Introduction to Data Visualization notes for spatial analysis.

Examples

summary(college_enrollment)

#>    school_id                                             name         label   
#>  Min.   : 1.00   Ball State University                     : 1           :28  
#>  1st Qu.:10.50   Bowling Green State University-Main Campus: 1   IUB     : 1  
#>  Median :20.00   Central Michigan University               : 1   MSU     : 1  
#>  Mean   :20.03   DePaul University                         : 1   OSU     : 1  
#>  3rd Qu.:29.50   Grand Valley State University             : 1   UChicago: 1  
#>  Max.   :40.00   Illinois State University                 : 1   UIowa   : 1  
#>                  (Other)                                   :33   (Other) : 6  
#>      state            city         lat             long         liberal_arts   
#>  IL     :9   Chicago    : 4   Min.   :37.14   Min.   :-93.65   Min.   :0.0000  
#>  MI     :6   Allendale  : 1   1st Qu.:39.64   1st Qu.:-89.32   1st Qu.:1.0000  
#>  OH     :6   Ames       : 1   Median :41.65   Median :-87.60   Median :1.0000  
#>  IN     :5   Ann Arbor  : 1   Mean   :41.09   Mean   :-87.41   Mean   :0.8718  
#>  MO     :4   Athens     : 1   3rd Qu.:42.28   3rd Qu.:-84.78   3rd Qu.:1.0000  
#>  WI     :4   Bloomington: 1   Max.   :44.98   Max.   :-81.34   Max.   :1.0000  
#>  (Other):5   (Other)    :30                                                    
#>    undergrad      enrollment   
#>  Min.   :1398   Min.   : 5781  
#>  1st Qu.:3167   1st Qu.:20102  
#>  Median :5284   Median :25072  
#>  Mean   :5117   Mean   :26785  
#>  3rd Qu.:6392   3rd Qu.:31186  
#>  Max.   :9560   Max.   :55014  
#>                                

table(college_enrollment$liberal_arts)

#> 
#>  0  1 
#>  5 34