Merge on State Function

mergeOnState merges two dataframe on state names. The columns containing state names in the two dataframes do not have to have the same format.

mergeOnState(df1, col1, df2, col2, merge = "left")

Arguments

df1
A dataframe containing a column with state names

col1
The column within df1 containing state names

df2
A dataframe containing a column with state names

col2
The column within df2 containing state names

merge
The type of merge ("left", "right", "inner", "full") that should be used to merge the two dataframes

Value

A dataframe

Examples

data(bingedrinking)
data(gdp)
mergeOnState(bingedrinking, LocationAbbr, gdp, X, merge = "inner")

#> Warning: DC is not a state

#> Warning: DC is not a state

#> Warning: DC is not a state

#> Warning: District of Columbia is not a state

#>     LocationAbbr BingeDrinking   Group              X gdpPerCap
#> 1             AL          13.8 Overall        Alabama     36750
#> 2             AK          17.8 Overall         Alaska     66835
#> 3             AZ          16.5 Overall        Arizona     38276
#> 4             AR          15.9 Overall       Arkansas     36259
#> 5             CA          16.5 Overall     California     56851
#> 6             CO          19.5 Overall       Colorado     52558
#> 7             CT          18.1 Overall    Connecticut     62800
#> 8             DE          18.0 Overall       Delaware     63783
#> 9             FL          17.0 Overall        Florida     39218
#> 10            GA          13.6 Overall        Georgia     43555
#> 11            HI          19.6 Overall         Hawaii     49945
#> 12            ID          16.9 Overall          Idaho     35455
#> 13            IL          20.7 Overall       Illinois     53669
#> 14            IN          18.5 Overall        Indiana     45118
#> 15            IA          22.5 Overall           Iowa     49532
#> 16            KS          16.8 Overall         Kansas     46132
#> 17            KY          15.6 Overall       Kentucky     38865
#> 18            LA          17.3 Overall      Louisiana     44587
#> 19            ME          21.1 Overall          Maine     38431
#> 20            MD          16.1 Overall       Maryland     54626
#> 21            MA          19.0 Overall  Massachusetts     64017
#> 22            MI          20.2 Overall       Michigan     42401
#> 23            MN          22.3 Overall      Minnesota     53562
#> 24            MS          12.8 Overall    Mississippi     31504
#> 25            MO          19.5 Overall       Missouri     43145
#> 26            MT          20.6 Overall        Montana     39046
#> 27            NE          21.1 Overall       Nebraska     52773
#> 28            NV          16.6 Overall         Nevada     43126
#> 29            NH          19.8 Overall  New Hampshire     50162
#> 30            NJ          17.2 Overall     New Jersey     56663
#> 31            NM          15.7 Overall     New Mexico     41529
#> 32            NY          18.2 Overall       New York     63390
#> 33            NC          15.5 Overall North Carolina     43680
#> 34            ND          25.2 Overall   North Dakota     67278
#> 35            OH          19.4 Overall           Ohio     47109
#> 36            OK          12.5 Overall       Oklahoma     45042
#> 37            OR          17.4 Overall         Oregon     50009
#> 38            PA          20.9 Overall   Pennsylvania     50540
#> 39            RI          16.7 Overall   Rhode Island     47413
#> 40            SC          18.0 Overall South Carolina     36620
#> 41            SD          20.7 Overall   South Dakota     47785
#> 42            TN          14.1 Overall      Tennessee     42647
#> 43            TX          17.9 Overall          Texas     54964
#> 44            UT          12.0 Overall           Utah     43688
#> 45            VT          20.3 Overall        Vermont     43122
#> 46            VA          16.3 Overall       Virginia     51540
#> 47            WA          17.0 Overall     Washington     55577
#> 48            WV          12.7 Overall  West Virginia     36817
#> 49            WI          26.0 Overall      Wisconsin     46893
#> 50            WY          19.1 Overall        Wyoming     61389
#> 51            AL          18.3    Male        Alabama     36750
#> 52            AK          20.9    Male         Alaska     66835
#> 53            AZ          22.5    Male        Arizona     38276
#> 54            AR          21.2    Male       Arkansas     36259
#> 55            CA          21.5    Male     California     56851
#> 56            CO          24.5    Male       Colorado     52558
#> 57            CT          22.9    Male    Connecticut     62800
#> 58            DE          21.6    Male       Delaware     63783
#> 59            FL          21.4    Male        Florida     39218
#> 60            GA          17.5    Male        Georgia     43555
#> 61            HI          24.7    Male         Hawaii     49945
#> 62            ID          21.3    Male          Idaho     35455
#> 63            IL          26.0    Male       Illinois     53669
#> 64            IN          23.9    Male        Indiana     45118
#> 65            IA          28.3    Male           Iowa     49532
#> 66            KS          22.5    Male         Kansas     46132
#> 67            KY          21.1    Male       Kentucky     38865
#> 68            LA          22.4    Male      Louisiana     44587
#> 69            ME          26.3    Male          Maine     38431
#> 70            MD          20.1    Male       Maryland     54626
#> 71            MA          23.4    Male  Massachusetts     64017
#> 72            MI          25.1    Male       Michigan     42401
#> 73            MN          27.3    Male      Minnesota     53562
#> 74            MS          18.5    Male    Mississippi     31504
#> 75            MO          24.6    Male       Missouri     43145
#> 76            MT          26.7    Male        Montana     39046
#> 77            NE          27.9    Male       Nebraska     52773
#> 78            NV          20.7    Male         Nevada     43126
#> 79            NH          24.9    Male  New Hampshire     50162
#> 80            NJ          21.3    Male     New Jersey     56663
#> 81            NM          21.1    Male     New Mexico     41529
#> 82            NY          23.1    Male       New York     63390
#> 83            NC          19.5    Male North Carolina     43680
#> 84            ND          32.0    Male   North Dakota     67278
#> 85            OH          24.8    Male           Ohio     47109
#> 86            OK          16.9    Male       Oklahoma     45042
#> 87            OR          21.9    Male         Oregon     50009
#> 88            PA          26.3    Male   Pennsylvania     50540
#> 89            RI          21.9    Male   Rhode Island     47413
#> 90            SC          23.8    Male South Carolina     36620
#> 91            SD          25.3    Male   South Dakota     47785
#> 92            TN          17.5    Male      Tennessee     42647
#> 93            TX          23.9    Male          Texas     54964
#> 94            UT          16.0    Male           Utah     43688
#> 95            VT          25.5    Male        Vermont     43122
#> 96            VA          20.3    Male       Virginia     51540
#> 97            WA          21.0    Male     Washington     55577
#> 98            WV          18.3    Male  West Virginia     36817
#> 99            WI          30.7    Male      Wisconsin     46893
#> 100           WY          24.6    Male        Wyoming     61389
#> 101           AL           9.7  Female        Alabama     36750
#> 102           AK          14.9  Female         Alaska     66835
#> 103           AZ          10.7  Female        Arizona     38276
#> 104           AR          10.8  Female       Arkansas     36259
#> 105           CA          11.5  Female     California     56851
#> 106           CO          14.5  Female       Colorado     52558
#> 107           CT          13.4  Female    Connecticut     62800
#> 108           DE          14.9  Female       Delaware     63783
#> 109           FL          12.8  Female        Florida     39218
#> 110           GA          10.0  Female        Georgia     43555
#> 111           HI          14.2  Female         Hawaii     49945
#> 112           ID          12.4  Female          Idaho     35455
#> 113           IL          15.4  Female       Illinois     53669
#> 114           IN          13.2  Female        Indiana     45118
#> 115           IA          16.7  Female           Iowa     49532
#> 116           KS          11.2  Female         Kansas     46132
#> 117           KY          10.3  Female       Kentucky     38865
#> 118           LA          12.5  Female      Louisiana     44587
#> 119           ME          16.0  Female          Maine     38431
#> 120           MD          12.2  Female       Maryland     54626
#> 121           MA          14.9  Female  Massachusetts     64017
#> 122           MI          15.5  Female       Michigan     42401
#> 123           MN          17.4  Female      Minnesota     53562
#> 124           MS           7.5  Female    Mississippi     31504
#> 125           MO          14.6  Female       Missouri     43145
#> 126           MT          14.4  Female        Montana     39046
#> 127           NE          14.2  Female       Nebraska     52773
#> 128           NV          12.4  Female         Nevada     43126
#> 129           NH          14.7  Female  New Hampshire     50162
#> 130           NJ          13.1  Female     New Jersey     56663
#> 131           NM          10.3  Female     New Mexico     41529
#> 132           NY          13.6  Female       New York     63390
#> 133           NC          11.7  Female North Carolina     43680
#> 134           ND          18.0  Female   North Dakota     67278
#> 135           OH          14.2  Female           Ohio     47109
#> 136           OK           8.3  Female       Oklahoma     45042
#> 137           OR          12.9  Female         Oregon     50009
#> 138           PA          15.7  Female   Pennsylvania     50540
#> 139           RI          11.8  Female   Rhode Island     47413
#> 140           SC          12.4  Female South Carolina     36620
#> 141           SD          16.0  Female   South Dakota     47785
#> 142           TN          10.8  Female      Tennessee     42647
#> 143           TX          11.9  Female          Texas     54964
#> 144           UT           8.0  Female           Utah     43688
#> 145           VT          15.4  Female        Vermont     43122
#> 146           VA          12.5  Female       Virginia     51540
#> 147           WA          13.0  Female     Washington     55577
#> 148           WV           7.2  Female  West Virginia     36817
#> 149           WI          21.2  Female      Wisconsin     46893
#> 150           WY          13.3  Female        Wyoming     61389